% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/addGrowthRate_function.R
\name{addGrowthRate}
\alias{addGrowthRate}
\title{Calculate a growth rate for a metric in a dataframe}
\usage{
addGrowthRate(data.in, metric, date.col = "date", growth.per = 364)
}
\arguments{
\item{data.in}{The dataframe containing a date and metric}

\item{metric}{}

\item{date.col}{The date column in the dataframe}

\item{growth.per}{}
}
\description{
This function takes a dataframe with a date column and a metric column,
along with any other dimensions in the dataset, and creates a new
metric column with the growth rate of the provided metric compared
to it's value some number of periods before (specified by growth.per)
}

