% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotMDTS_function.R
\name{addMissingZeros}
\alias{addMissingZeros}
\title{Replace missing data points for time series with zeros}
\usage{
addMissingZeros(data.in)
}
\description{
This function takes a multi-dimensional data frame and passes
back an identical one but with rows added for dates missing
values for each unique combination of dimensions in the dataframe
that has existing values (e.g. it won't complete for already
empty combinations).
}

