% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/addTimeShift_function.R
\name{addTimeShift}
\alias{addTimeShift}
\title{Add column of time-shifted shifted metrics}
\usage{
addTimeShift(data.in, metric, date.col = "date", time.shift = 1,
  shift.per = 364)
}
\arguments{
\item{data.in}{The dataframe containing a date and metric}

\item{metric}{}

\item{date.col}{The date column in the dataframe}

\item{time.shift}{The number of period shifts to apply}

\item{shift.per}{The period of time to shift by}
}
\description{
This function takes a dataframe with a date column and a metric column,
along with any other dimensions in the dataset, and creates a new
metric column time-shifted by the parameters specified. E.g. clicks for
this row one year ago or one week ago.
}

