% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/aggregateByDate_function.R
\name{aggregateByDate}
\alias{aggregateByDate}
\title{Aggregate a dataframe by date dimensions}
\usage{
aggregateByDate(data.in, date.col = "date", by.date = "day")
}
\arguments{
\item{data.in}{A dataframe with a date column (and other metrics or dimensions)}

\item{date.col}{The date column to use (defaults to "date")}

\item{by.date}{The granularity to summarize by: day, week, month, year}
}
\description{
This function takes a dataframe with a date dimension and aggregates
it based on common date conventions -- daily, weekly, monthly, quarterly,
yearly. It will return a dataframe with the same dimensionality and column
names, but the "date" column will be aggregated with all the values of the
period on the first date of that period (e.g. yearly will return as yyyy-01-01)
}

