% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotMDTS_function.R
\name{plotMDTS}
\alias{plotMDTS}
\title{Plot a Multi-Dimensional Time-Series}
\usage{
plotMDTS(data.in, metric, filters = NULL, group = NULL, facet = NULL,
  by.date = "day", time.shift = NULL, shift.per = 364, growth.per = 364,
  title = metric, ylab = metric, xlab = "Date", add.weekends = FALSE,
  add.zeros = FALSE, y.zero = FALSE)
}
\arguments{
\item{data.in}{A time-series data frame with a column of dates named "date"}

\item{metric}{The metric you want to plot. Count unique values of a column with: "n(my_dimension)"}
}
\description{
This wrapper enables you to quickly and powerfully plot a multi-dimensional
time-series dataset using ggplot2 and dplyr.
}
\examples{
plotMDTS(example_data, metric="raisedAmt", filters = ~date > '2007-01-01',
                   group = "category", facet = "state")
}

